requirements.txt
aenum==2.2.6
APScheduler==3.6.3
attrs==21.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
bot-vk==1.7
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.8
DateTime==4.3
fbchat==1.9.7
idna==3.3
iniconfig==1.1.1
lxml==4.6.4
packaging==21.3
paho-mqtt==1.6.1
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.6
pytest==6.2.5
python-telegram-bot==13.8.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
six==1.16.0
skype-chatbot==1.0.4
soupsieve==2.3.1
toml==0.10.2
tornado==6.1
transitions==0.8.10
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
vk-api==11.9.6
zope.interface==5.4.0
